+++ src/main/java/com/mekeng/github/proxy/ClientProxy.java
@@ -7,11 +7,18 @@ import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 import appeng.api.client.AEKeyRendering;
 import appeng.api.integrations.hei.IngredientConverters;
 
+import com.mekeng.github.client.ClientRegistryHandler;
 import com.mekeng.github.client.hei.GasIngredientConverter;
 import com.mekeng.github.client.render.GasKeyRenderHandler;
+import com.mekeng.github.common.RegistryHandler;
 import com.mekeng.github.common.me.AEGasKeyType;
 
 public class ClientProxy extends CommonProxy {
+
+    @Override
+    public RegistryHandler createRegistryHandler() {
+        return new ClientRegistryHandler();
+    }
 
     @Override
     public void preInit(FMLPreInitializationEvent event) {